LogSend RPU for gui
