+++ go/protocol/keybase1/config.go
@@ -400,6 +400,12 @@ func (o FullStatus) DeepCopy() FullStatus {
 	}
 }
 
+type LogSendID string
+
+func (o LogSendID) DeepCopy() LogSendID {
+	return o
+}
+
 type AllProvisionedUsernames struct {
 	DefaultUsername      string   `codec:"defaultUsername" json:"defaultUsername"`
 	ProvisionedUsernames []string `codec:"provisionedUsernames" json:"provisionedUsernames"`
@@ -775,6 +781,13 @@ type GetFullStatusArg struct {
 	SessionID int `codec:"sessionID" json:"sessionID"`
 }
 
+type LogSendArg struct {
+	SessionID  int    `codec:"sessionID" json:"sessionID"`
+	StatusJSON string `codec:"statusJSON" json:"statusJSON"`
+	Feedback   string `codec:"feedback" json:"feedback"`
+	SendLogs   bool   `codec:"sendLogs" json:"sendLogs"`
+}
+
 type GetAllProvisionedUsernamesArg struct {
 	SessionID int `codec:"sessionID" json:"sessionID"`
 }
@@ -848,6 +861,7 @@ type ConfigInterface interface {
 	GetCurrentStatus(context.Context, int) (CurrentStatus, error)
 	GetClientStatus(context.Context, int) ([]ClientStatus, error)
 	GetFullStatus(context.Context, int) (*FullStatus, error)
+	LogSend(context.Context, LogSendArg) (LogSendID, error)
 	GetAllProvisionedUsernames(context.Context, int) (AllProvisionedUsernames, error)
 	GetConfig(context.Context, int) (Config, error)
 	// Change user config.
@@ -919,6 +933,21 @@ func ConfigProtocol(i ConfigInterface) rpc.Protocol {
 						return
 					}
 					ret, err = i.GetFullStatus(ctx, typedArgs[0].SessionID)
+					return
+				},
+			},
+			"logSend": {
+				MakeArg: func() interface{} {
+					var ret [1]LogSendArg
+					return &ret
+				},
+				Handler: func(ctx context.Context, args interface{}) (ret interface{}, err error) {
+					typedArgs, ok := args.(*[1]LogSendArg)
+					if !ok {
+						err = rpc.NewTypeError((*[1]LogSendArg)(nil), args)
+						return
+					}
+					ret, err = i.LogSend(ctx, typedArgs[0])
 					return
 				},
 			},
@@ -1170,6 +1199,11 @@ func (c ConfigClient) GetClientStatus(ctx context.Context, sessionID int) (res [
 func (c ConfigClient) GetFullStatus(ctx context.Context, sessionID int) (res *FullStatus, err error) {
 	__arg := GetFullStatusArg{SessionID: sessionID}
 	err = c.Cli.Call(ctx, "keybase.1.config.getFullStatus", []interface{}{__arg}, &res)
+	return
+}
+
+func (c ConfigClient) LogSend(ctx context.Context, __arg LogSendArg) (res LogSendID, err error) {
+	err = c.Cli.Call(ctx, "keybase.1.config.logSend", []interface{}{__arg}, &res)
 	return
 }
