+++ protocol/json/keybase1/config.json
@@ -467,6 +467,12 @@
     },
     {
       "type": "record",
+      "name": "LogSendID",
+      "fields": [],
+      "typedef": "string"
+    },
+    {
+      "type": "record",
       "name": "AllProvisionedUsernames",
       "fields": [
         {
@@ -778,6 +784,27 @@
         null,
         "FullStatus"
       ]
+    },
+    "logSend": {
+      "request": [
+        {
+          "name": "sessionID",
+          "type": "int"
+        },
+        {
+          "name": "statusJSON",
+          "type": "string"
+        },
+        {
+          "name": "feedback",
+          "type": "string"
+        },
+        {
+          "name": "sendLogs",
+          "type": "boolean"
+        }
+      ],
+      "response": "LogSendID"
     },
     "getAllProvisionedUsernames": {
       "request": [
